App and ImageFile classes' methods updated to allow for a filechooser openFile() method, updateD() method to keep display updated, and start() code reduced to essentially just calling update() on startup
